added enhanced state with all the enhanced features like undo, ..
