+++ packages/core/src/collection/group/index.ts
@@ -1,5 +1,5 @@
 import {
-  PersistableState,
+  EnhancedState,
   Collection,
   DefaultItem,
   ItemKey,
@@ -23,7 +23,7 @@ import {
 export class Group<
   DataType extends Object = DefaultItem,
   ValueType = Array<ItemKey> // To extract the Group Type Value in Integration methods like 'useAgile()'
-> extends PersistableState<Array<ItemKey>> {
+> extends EnhancedState<Array<ItemKey>> {
   // Collection the Group belongs to
   collection: () => Collection<DataType>;
