+++ packages/core/src/internal.ts
@@ -38,7 +38,7 @@ export * from './state';
 export * from './state/state.observer';
 export * from './state/state.persistent';
 export * from './state/state.runtime.job';
-export * from './state/persistableState';
+export * from './state/state.enhanced';
 
 // Computed
 export * from './computed';
