+++ packages/core/src/state/index.ts
@@ -5,7 +5,7 @@ import {
   removeProperties,
   CreateAgileSubInstanceInterface,
   shared,
-  PersistableState,
+  EnhancedState,
 } from '../internal';
 
 export * from './state';
@@ -32,7 +32,7 @@ export interface CreateStateConfigInterfaceWithAgile
  * @param initialValue - Initial value of the State.
  * @param config - Configuration object
  */
-export function createState<ValueType = any>(
+export function createLightState<ValueType = any>(
   initialValue: ValueType,
   config: CreateStateConfigInterfaceWithAgile = {}
 ): State<ValueType> {
@@ -46,14 +46,29 @@ export function createState<ValueType = any>(
   );
 }
 
-export function createPersistableState<ValueType = any>(
+/**
+ * Returns a newly created State.
+ *
+ * A State manages a piece of Information
+ * that we need to remember globally at a later point in time.
+ * While providing a toolkit to use and mutate this piece of Information.
+ *
+ * You can create as many global States as you need.
+ *
+ * [Learn more..](https://agile-ts.org/docs/core/agile-instance/methods#createstate)
+ *
+ * @public
+ * @param initialValue - Initial value of the State.
+ * @param config - Configuration object
+ */
+export function createState<ValueType = any>(
   initialValue: ValueType,
   config: CreateStateConfigInterfaceWithAgile = {}
-): PersistableState<ValueType> {
+): EnhancedState<ValueType> {
   config = defineConfig(config, {
     agileInstance: shared,
   });
-  return new PersistableState<ValueType>(
+  return new EnhancedState<ValueType>(
     config.agileInstance as any,
     initialValue,
     removeProperties(config, ['agileInstance'])
